[PAL/Linux-SGX] Split `host_enclave.[c|h]` into two
Previously, `host_enclave.[c|h]` contained both ECALLs and OCALLs. This
commit splits the file into `host_ecalls.[c|h]` and `host_ocalls.c`.
Also, this commit removes useless ODEBUG/EDEBUG macros.
