+++ pal/src/host/linux-sgx/host_ecalls.h
