+++ pal/src/host/linux-sgx/host_ocalls.c
@@ -1,7 +1,7 @@
 /* SPDX-License-Identifier: LGPL-3.0-or-later */
-/* Copyright (C) 2014 Stony Brook University
- */
-#include "host_enclave.h"
+/* Copyright (C) 2014 Stony Brook University */
+
+#include <stddef.h> /* must be included before linux/signal.h */
 
 #include <asm/errno.h>
 #include <asm/ioctls.h>
@@ -15,26 +15,25 @@
 
 #include "cpu.h"
 #include "debug_map.h"
+#include "host_ecalls.h"
 #include "host_internal.h"
 #include "host_process.h"
 #include "linux_utils.h"
-#include "pal_ecall_types.h"
 #include "pal_ocall_types.h"
 #include "pal_rpc_queue.h"
 #include "pal_tls.h"
+#include "pal_topology.h"
+#include "sgx_arch.h"
 #include "sigset.h"
 
 #define DEFAULT_BACKLOG 2048
 
-#define ODEBUG(code, ms) \
-    do {                 \
-    } while (0)
-
 extern bool g_vtune_profile_enabled;
+
+rpc_queue_t* g_rpc_queue = NULL; /* pointer to untrusted queue */
 
 static long sgx_ocall_exit(void* pms) {
     ms_ocall_exit_t* ms = (ms_ocall_exit_t*)pms;
-    ODEBUG(OCALL_EXIT, NULL);
 
     if (ms->ms_exitcode != (int)((uint8_t)ms->ms_exitcode)) {
         log_debug("Saturation error in exit code %d, getting rounded down to %u",
@@ -88,7 +87,6 @@ static long sgx_ocall_exit(void* pms) {
 static long sgx_ocall_mmap_untrusted(void* pms) {
     ms_ocall_mmap_untrusted_t* ms = (ms_ocall_mmap_untrusted_t*)pms;
     void* addr;
-    ODEBUG(OCALL_MMAP_UNTRUSTED, ms);
 
     addr = (void*)DO_SYSCALL(mmap, ms->ms_addr, ms->ms_size, ms->ms_prot, ms->ms_flags, ms->ms_fd,
                              ms->ms_offset);
@@ -101,14 +99,12 @@ static long sgx_ocall_mmap_untrusted(void* pms) {
 
 static long sgx_ocall_munmap_untrusted(void* pms) {
     ms_ocall_munmap_untrusted_t* ms = (ms_ocall_munmap_untrusted_t*)pms;
-    ODEBUG(OCALL_MUNMAP_UNTRUSTED, ms);
     DO_SYSCALL(munmap, ms->ms_addr, ms->ms_size);
     return 0;
 }
 
 static long sgx_ocall_cpuid(void* pms) {
     ms_ocall_cpuid_t* ms = (ms_ocall_cpuid_t*)pms;
-    ODEBUG(OCALL_CPUID, ms);
     __asm__ volatile("cpuid"
                      : "=a"(ms->ms_values[0]),
                        "=b"(ms->ms_values[1]),
@@ -122,7 +118,6 @@ static long sgx_ocall_cpuid(void* pms) {
 static long sgx_ocall_open(void* pms) {
     ms_ocall_open_t* ms = (ms_ocall_open_t*)pms;
     long ret;
-    ODEBUG(OCALL_OPEN, ms);
     // FIXME: No idea why someone hardcoded O_CLOEXEC here. We should drop it and carefully
     // investigate if this cause any descriptor leaks.
     ret = DO_SYSCALL_INTERRUPTIBLE(open, ms->ms_pathname, ms->ms_flags | O_CLOEXEC, ms->ms_mode);
@@ -131,7 +126,6 @@ static long sgx_ocall_open(void* pms) {
 
 static long sgx_ocall_close(void* pms) {
     ms_ocall_close_t* ms = (ms_ocall_close_t*)pms;
-    ODEBUG(OCALL_CLOSE, ms);
     /* Callers cannot retry close on `-EINTR`, so we do not call `DO_SYSCALL_INTERRUPTIBLE`. */
     return DO_SYSCALL(close, ms->ms_fd);
 }
@@ -139,7 +133,6 @@ static long sgx_ocall_close(void* pms) {
 static long sgx_ocall_read(void* pms) {
     ms_ocall_read_t* ms = (ms_ocall_read_t*)pms;
     long ret;
-    ODEBUG(OCALL_READ, ms);
     ret = DO_SYSCALL_INTERRUPTIBLE(read, ms->ms_fd, ms->ms_buf, ms->ms_count);
     return ret;
 }
@@ -147,7 +140,6 @@ static long sgx_ocall_read(void* pms) {
 static long sgx_ocall_write(void* pms) {
     ms_ocall_write_t* ms = (ms_ocall_write_t*)pms;
     long ret;
-    ODEBUG(OCALL_WRITE, ms);
     ret = DO_SYSCALL_INTERRUPTIBLE(write, ms->ms_fd, ms->ms_buf, ms->ms_count);
     return ret;
 }
@@ -155,7 +147,6 @@ static long sgx_ocall_write(void* pms) {
 static long sgx_ocall_pread(void* pms) {
     ms_ocall_pread_t* ms = (ms_ocall_pread_t*)pms;
     long ret;
-    ODEBUG(OCALL_PREAD, ms);
     ret = DO_SYSCALL_INTERRUPTIBLE(pread64, ms->ms_fd, ms->ms_buf, ms->ms_count, ms->ms_offset);
     return ret;
 }
@@ -163,7 +154,6 @@ static long sgx_ocall_pread(void* pms) {
 static long sgx_ocall_pwrite(void* pms) {
     ms_ocall_pwrite_t* ms = (ms_ocall_pwrite_t*)pms;
     long ret;
-    ODEBUG(OCALL_PWRITE, ms);
     ret = DO_SYSCALL_INTERRUPTIBLE(pwrite64, ms->ms_fd, ms->ms_buf, ms->ms_count, ms->ms_offset);
     return ret;
 }
@@ -171,7 +161,6 @@ static long sgx_ocall_pwrite(void* pms) {
 static long sgx_ocall_fstat(void* pms) {
     ms_ocall_fstat_t* ms = (ms_ocall_fstat_t*)pms;
     long ret;
-    ODEBUG(OCALL_FSTAT, ms);
     ret = DO_SYSCALL_INTERRUPTIBLE(fstat, ms->ms_fd, &ms->ms_stat);
     return ret;
 }
@@ -180,7 +169,6 @@ static long sgx_ocall_fionread(void* pms) {
     ms_ocall_fionread_t* ms = (ms_ocall_fionread_t*)pms;
     long ret;
     int val;
-    ODEBUG(OCALL_FIONREAD, ms);
     ret = DO_SYSCALL_INTERRUPTIBLE(ioctl, ms->ms_fd, FIONREAD, &val);
     return ret < 0 ? ret : val;
 }
@@ -189,7 +177,6 @@ static long sgx_ocall_fsetnonblock(void* pms) {
     ms_ocall_fsetnonblock_t* ms = (ms_ocall_fsetnonblock_t*)pms;
     long ret;
     int flags;
-    ODEBUG(OCALL_FSETNONBLOCK, ms);
 
     ret = DO_SYSCALL(fcntl, ms->ms_fd, F_GETFL);
     if (ret < 0)
@@ -210,21 +197,18 @@ static long sgx_ocall_fsetnonblock(void* pms) {
 static long sgx_ocall_fchmod(void* pms) {
     ms_ocall_fchmod_t* ms = (ms_ocall_fchmod_t*)pms;
     long ret;
-    ODEBUG(OCALL_FCHMOD, ms);
     ret = DO_SYSCALL(fchmod, ms->ms_fd, ms->ms_mode);
     return ret;
 }
 
 static long sgx_ocall_fsync(void* pms) {
     ms_ocall_fsync_t* ms = (ms_ocall_fsync_t*)pms;
-    ODEBUG(OCALL_FSYNC, ms);
     return DO_SYSCALL_INTERRUPTIBLE(fsync, ms->ms_fd);
 }
 
 static long sgx_ocall_ftruncate(void* pms) {
     ms_ocall_ftruncate_t* ms = (ms_ocall_ftruncate_t*)pms;
     long ret;
-    ODEBUG(OCALL_FTRUNCATE, ms);
     ret = DO_SYSCALL(ftruncate, ms->ms_fd, ms->ms_length);
     return ret;
 }
@@ -232,7 +216,6 @@ static long sgx_ocall_ftruncate(void* pms) {
 static long sgx_ocall_mkdir(void* pms) {
     ms_ocall_mkdir_t* ms = (ms_ocall_mkdir_t*)pms;
     long ret;
-    ODEBUG(OCALL_MKDIR, ms);
     ret = DO_SYSCALL(mkdir, ms->ms_pathname, ms->ms_mode);
     return ret;
 }
@@ -240,14 +223,12 @@ static long sgx_ocall_mkdir(void* pms) {
 static long sgx_ocall_getdents(void* pms) {
     ms_ocall_getdents_t* ms = (ms_ocall_getdents_t*)pms;
     long ret;
-    ODEBUG(OCALL_GETDENTS, ms);
     unsigned int count = ms->ms_size <= UINT_MAX ? ms->ms_size : UINT_MAX;
     ret = DO_SYSCALL_INTERRUPTIBLE(getdents64, ms->ms_fd, ms->ms_dirp, count);
     return ret;
 }
 
 static long sgx_ocall_resume_thread(void* pms) {
-    ODEBUG(OCALL_RESUME_THREAD, pms);
     int tid = get_tid_from_tcs(pms);
     if (tid < 0)
         return tid;
@@ -258,7 +239,6 @@ static long sgx_ocall_resume_thread(void* pms) {
 
 static long sgx_ocall_sched_setaffinity(void* pms) {
     ms_ocall_sched_setaffinity_t* ms = (ms_ocall_sched_setaffinity_t*)pms;
-    ODEBUG(OCALL_SCHED_SETAFFINITY, ms);
     int tid = get_tid_from_tcs(ms->ms_tcs);
     if (tid < 0)
         return tid;
@@ -269,7 +249,6 @@ static long sgx_ocall_sched_setaffinity(void* pms) {
 
 static long sgx_ocall_sched_getaffinity(void* pms) {
     ms_ocall_sched_getaffinity_t* ms = (ms_ocall_sched_getaffinity_t*)pms;
-    ODEBUG(OCALL_SCHED_GETAFFINITY, ms);
     int tid = get_tid_from_tcs(ms->ms_tcs);
     if (tid < 0)
         return tid;
@@ -280,13 +259,11 @@ static long sgx_ocall_sched_getaffinity(void* pms) {
 
 static long sgx_ocall_clone_thread(void* pms) {
     __UNUSED(pms);
-    ODEBUG(OCALL_CLONE_THREAD, pms);
     return clone_thread();
 }
 
 static long sgx_ocall_create_process(void* pms) {
     ms_ocall_create_process_t* ms = (ms_ocall_create_process_t*)pms;
-    ODEBUG(OCALL_CREATE_PROCESS, ms);
 
     return sgx_create_process(ms->ms_nargs, ms->ms_args, g_pal_enclave.raw_manifest_data,
                               &ms->ms_stream_fd);
@@ -295,7 +272,6 @@ static long sgx_ocall_create_process(void* pms) {
 static long sgx_ocall_futex(void* pms) {
     ms_ocall_futex_t* ms = (ms_ocall_futex_t*)pms;
     long ret;
-    ODEBUG(OCALL_FUTEX, ms);
 
     struct timespec timeout = { 0 };
     bool have_timeout = ms->ms_timeout_us != (uint64_t)-1;
@@ -380,7 +356,6 @@ static long sgx_ocall_listen(void* pms) {
     ms_ocall_listen_t* ms = (ms_ocall_listen_t*)pms;
     long ret;
     int fd;
-    ODEBUG(OCALL_LISTEN, ms);
 
     if (ms->ms_addrlen > INT_MAX) {
         ret = -EINVAL;
@@ -436,7 +411,6 @@ err:
 static long sgx_ocall_accept(void* pms) {
     ms_ocall_accept_t* ms = (ms_ocall_accept_t*)pms;
     long ret;
-    ODEBUG(OCALL_ACCEPT, ms);
 
     if (ms->ms_addrlen > INT_MAX) {
         ret = -EINVAL;
@@ -462,7 +436,6 @@ static long sgx_ocall_connect(void* pms) {
     ms_ocall_connect_t* ms = (ms_ocall_connect_t*)pms;
     long ret;
     int fd;
-    ODEBUG(OCALL_CONNECT, ms);
 
     if (ms->ms_addrlen > INT_MAX || ms->ms_bind_addrlen > INT_MAX) {
         ret = -EINVAL;
@@ -565,7 +538,6 @@ static long sgx_ocall_connect_simple(void* pms) {
 static long sgx_ocall_recv(void* pms) {
     ms_ocall_recv_t* ms = (ms_ocall_recv_t*)pms;
     long ret;
-    ODEBUG(OCALL_RECV, ms);
 
     if (ms->ms_addr && ms->ms_addrlen > INT_MAX) {
         return -EINVAL;
@@ -603,7 +575,6 @@ static long sgx_ocall_recv(void* pms) {
 static long sgx_ocall_send(void* pms) {
     ms_ocall_send_t* ms = (ms_ocall_send_t*)pms;
     long ret;
-    ODEBUG(OCALL_SEND, ms);
 
     if (ms->ms_addr && ms->ms_addrlen > INT_MAX) {
         return -EINVAL;
@@ -630,7 +601,6 @@ static long sgx_ocall_send(void* pms) {
 static long sgx_ocall_setsockopt(void* pms) {
     ms_ocall_setsockopt_t* ms = (ms_ocall_setsockopt_t*)pms;
     long ret;
-    ODEBUG(OCALL_SETSOCKOPT, ms);
     if (ms->ms_optlen > INT_MAX) {
         return -EINVAL;
     }
@@ -641,14 +611,12 @@ static long sgx_ocall_setsockopt(void* pms) {
 
 static long sgx_ocall_shutdown(void* pms) {
     ms_ocall_shutdown_t* ms = (ms_ocall_shutdown_t*)pms;
-    ODEBUG(OCALL_SHUTDOWN, ms);
     DO_SYSCALL_INTERRUPTIBLE(shutdown, ms->ms_sockfd, ms->ms_how);
     return 0;
 }
 
 static long sgx_ocall_gettime(void* pms) {
     ms_ocall_gettime_t* ms = (ms_ocall_gettime_t*)pms;
-    ODEBUG(OCALL_GETTIME, ms);
     struct timeval tv;
     DO_SYSCALL(gettimeofday, &tv, NULL);
     ms->ms_microsec = tv.tv_sec * (uint64_t)1000000 + tv.tv_usec;
@@ -657,7 +625,6 @@ static long sgx_ocall_gettime(void* pms) {
 
 static long sgx_ocall_sched_yield(void* pms) {
     __UNUSED(pms);
-    ODEBUG(OCALL_SCHED_YIELD, pms);
     DO_SYSCALL_INTERRUPTIBLE(sched_yield);
     return 0;
 }
@@ -665,7 +632,6 @@ static long sgx_ocall_sched_yield(void* pms) {
 static long sgx_ocall_poll(void* pms) {
     ms_ocall_poll_t* ms = (ms_ocall_poll_t*)pms;
     long ret;
-    ODEBUG(OCALL_POLL, ms);
 
     struct timespec* timeout = NULL;
     struct timespec end_time = { 0 };
@@ -695,7 +661,6 @@ static long sgx_ocall_poll(void* pms) {
 static long sgx_ocall_rename(void* pms) {
     ms_ocall_rename_t* ms = (ms_ocall_rename_t*)pms;
     long ret;
-    ODEBUG(OCALL_RENAME, ms);
     ret = DO_SYSCALL(rename, ms->ms_oldpath, ms->ms_newpath);
     return ret;
 }
@@ -703,7 +668,6 @@ static long sgx_ocall_rename(void* pms) {
 static long sgx_ocall_delete(void* pms) {
     ms_ocall_delete_t* ms = (ms_ocall_delete_t*)pms;
     long ret;
-    ODEBUG(OCALL_DELETE, ms);
 
     ret = DO_SYSCALL(unlink, ms->ms_pathname);
 
@@ -716,7 +680,6 @@ static long sgx_ocall_delete(void* pms) {
 static long sgx_ocall_eventfd(void* pms) {
     ms_ocall_eventfd_t* ms = (ms_ocall_eventfd_t*)pms;
     long ret;
-    ODEBUG(OCALL_EVENTFD, ms);
 
     ret = DO_SYSCALL(eventfd2, 0, ms->ms_flags);
 
@@ -764,7 +727,6 @@ static long sgx_ocall_debug_describe_location(void* pms) {
 
 static long sgx_ocall_get_quote(void* pms) {
     ms_ocall_get_quote_t* ms = (ms_ocall_get_quote_t*)pms;
-    ODEBUG(OCALL_GET_QUOTE, ms);
     return retrieve_quote(ms->ms_is_epid ? &ms->ms_spid : NULL, ms->ms_linkable, &ms->ms_report,
                           &ms->ms_nonce, &ms->ms_quote, &ms->ms_quote_len);
 }
@@ -816,12 +778,6 @@ sgx_ocall_fn_t ocall_table[OCALL_NR] = {
     [OCALL_EVENTFD]                  = sgx_ocall_eventfd,
     [OCALL_GET_QUOTE]                = sgx_ocall_get_quote,
 };
-
-#define EDEBUG(code, ms) \
-    do {                 \
-    } while (0)
-
-rpc_queue_t* g_rpc_queue = NULL; /* pointer to untrusted queue */
 
 static int rpc_thread_loop(void* arg) {
     __UNUSED(arg);
@@ -885,7 +841,7 @@ static int rpc_thread_loop(void* arg) {
     return 0;
 }
 
-static int start_rpc(size_t threads_cnt) {
+int start_rpc(size_t threads_cnt) {
     g_rpc_queue = (rpc_queue_t*)DO_SYSCALL(mmap, NULL,
                                            ALIGN_UP(sizeof(rpc_queue_t), PRESET_PAGESIZE),
                                            PROT_READ | PROT_WRITE, MAP_ANONYMOUS | MAP_PRIVATE,
@@ -929,43 +885,4 @@ static int start_rpc(size_t threads_cnt) {
     }
 
     return 0;
-}
-
-int ecall_enclave_start(char* libpal_uri, char* args, size_t args_size, char* env,
-                        size_t env_size, int parent_stream_fd, sgx_target_info_t* qe_targetinfo,
-                        struct pal_topo_info* topo_info) {
-    g_rpc_queue = NULL;
-
-    if (g_pal_enclave.rpc_thread_num > 0) {
-        int ret = start_rpc(g_pal_enclave.rpc_thread_num);
-        if (ret < 0) {
-            /* failed to create RPC threads */
-            return ret;
-        }
-        /* after this point, g_rpc_queue != NULL */
-    }
-
-    ms_ecall_enclave_start_t ms;
-    ms.ms_libpal_uri       = libpal_uri;
-    ms.ms_libpal_uri_len   = strlen(ms.ms_libpal_uri);
-    ms.ms_args             = args;
-    ms.ms_args_size        = args_size;
-    ms.ms_env              = env;
-    ms.ms_env_size         = env_size;
-    ms.ms_parent_stream_fd = parent_stream_fd;
-    ms.ms_qe_targetinfo    = qe_targetinfo;
-    ms.ms_topo_info        = topo_info;
-    ms.rpc_queue           = g_rpc_queue;
-    EDEBUG(ECALL_ENCLAVE_START, &ms);
-    return sgx_ecall(ECALL_ENCLAVE_START, &ms);
-}
-
-int ecall_thread_start(void) {
-    EDEBUG(ECALL_THREAD_START, NULL);
-    return sgx_ecall(ECALL_THREAD_START, NULL);
-}
-
-int ecall_thread_reset(void) {
-    EDEBUG(ECALL_THREAD_RESET, NULL);
-    return sgx_ecall(ECALL_THREAD_RESET, NULL);
 }
