+++ pal/src/host/linux-sgx/meson.build
@@ -127,12 +127,13 @@ libpal_sgx = shared_library('pal',
 # host (untrusted runtime)
 libpal_sgx_host = executable('loader',
     'common_manifest_sgx_parser.c',
-    'host_enclave.c',
+    'host_ecalls.c',
     'host_entry.S',
     'host_exception.c',
     'host_framework.c',
     'host_log.c',
     'host_main.c',
+    'host_ocalls.c',
     'host_perf_data.c',
     'host_platform.c',
     'host_process.c',
